edit tissus

+++ frontend/src/pages/admin/TissuListPage.js
@@ -14,9 +14,9 @@ import { Store } from '../../Store';
 import { getError } from '../../utils';
 import LoadingBox from '../../components/LoadingBox';
 import MessageBox from '../../components/MessageBox';
-import { faTrash } from '@fortawesome/pro-solid-svg-icons';
+import { faPlus, faTrash } from '@fortawesome/pro-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Col, Container, Form, Row, Table } from 'react-bootstrap';
+import { Col, Container, Form, Image, Row, Table } from 'react-bootstrap';
 import AdminMenu from '../../components/AdminMenu';
 import AdminCanvasMenu from '../../components/AdminCanvasMenu';
 
@@ -25,6 +25,7 @@ import 'jquery/dist/jquery.min.js';
 import 'datatables.net-dt/js/dataTables.dataTables';
 import 'datatables.net-dt/css/jquery.dataTables.min.css';
 import $ from 'jquery';
+import ModalTissuEdit from '../../components/ModalEditTissu';
 
 const reducer = (state, action) => {
   switch (action.type) {
@@ -46,6 +47,12 @@ const reducer = (state, action) => {
       return { ...state, loading: false };
     case 'ADD_RESET':
       return { ...state, loading: false, successAdd: false };
+    case 'UPDATE_REQUEST':
+      return { ...state, loadingUpdate: true };
+    case 'UPDATE_SUCCESS':
+      return { ...state, loadingUpdate: false };
+    case 'UPDATE_FAIL':
+      return { ...state, loadingUpdate: false };
     case 'DELETE_REQUEST':
       return { ...state, loading: true, successDelete: false };
     case 'DELETE_SUCCESS':
@@ -58,6 +65,17 @@ const reducer = (state, action) => {
       return { ...state, loading: false };
     case 'DELETE_RESET':
       return { ...state, loading: false, successDelete: false };
+    case 'UPLOAD_REQUEST':
+      return { ...state, loadingUpload: true, errorUpload: '' };
+    case 'UPLOAD_SUCCESS':
+      return {
+        ...state,
+        loadingUpload: false,
+        errorUpload: '',
+      };
+    case 'UPLOAD_FAIL':
+      return { ...state, loadingUpload: false, errorUpload: action.payload };
+
     default:
       return state;
   }
@@ -65,11 +83,13 @@ const reducer = (state, action) => {
 export default function TissuListPage() {
   const navigate = useNavigate();
 
-  const [{ loading, error, tissus, successDelete, successAdd }, dispatch] =
-    useReducer(reducer, {
-      loading: true,
-      error: '',
-    });
+  const [
+    { loading, loadingUpload, error, tissus, successDelete, successAdd },
+    dispatch,
+  ] = useReducer(reducer, {
+    loading: true,
+    error: '',
+  });
 
   const { state } = useContext(Store);
   const { userInfo } = state;
@@ -77,6 +97,9 @@ export default function TissuListPage() {
   $.DataTable = require('datatables.net');
   const tableRef = useRef();
   const [name, setName] = useState('');
+
+  const imageInputRef = useRef();
+  const [image, setImage] = useState('');
 
   useEffect(() => {
     const fetchData = async () => {
@@ -120,6 +143,7 @@ export default function TissuListPage() {
         `/api/tissus/add`,
         {
           name,
+          image,
         },
         {
           headers: { Authorization: `Bearer ${userInfo.token}` },
@@ -127,6 +151,7 @@ export default function TissuListPage() {
       );
       dispatch({ type: 'ADD_SUCCESS' });
       setName('');
+      if (imageInputRef) imageInputRef.current.value = null;
       toast.success('Tissu ajouté');
       navigate('/admin/tissus');
     } catch (err) {
@@ -135,18 +160,41 @@ export default function TissuListPage() {
     }
   };
 
-  const deleteHandler = async (tissu) => {
+  const uploadFileHandler = async (e, forImages) => {
+    const file = e.target.files[0];
+    const bodyFormData = new FormData();
+    bodyFormData.append('file', file);
     try {
-      dispatch({ type: 'DELETE_REQUEST' });
-      await axios.delete(`/api/tissus/${tissu._id}`, {
-        headers: { Authorization: `Bearer ${userInfo.token}` },
+      dispatch({ type: 'UPLOAD_REQUEST' });
+      const { data } = await axios.post('/api/upload/tissu', bodyFormData, {
+        headers: {
+          'Content-Type': 'multipart/form-data',
+          authorization: `Bearer ${userInfo.token}`,
+        },
       });
-      dispatch({ type: 'DELETE_SUCCESS' });
+      dispatch({ type: 'UPLOAD_SUCCESS' });
+      setImage(data.secure_url);
     } catch (err) {
-      toast.error(getError(error));
-      dispatch({
-        type: 'DELETE_FAIL',
-      });
+      toast.error(getError(err));
+      dispatch({ type: 'UPLOAD_FAIL', payload: getError(err) });
+    }
+  };
+
+  const deleteHandler = async (tissu) => {
+    if (window.confirm('Confirmer ?')) {
+      try {
+        dispatch({ type: 'DELETE_REQUEST' });
+        await axios.delete(`/api/tissus/${tissu._id}`, {
+          headers: { Authorization: `Bearer ${userInfo.token}` },
+        });
+
+        dispatch({ type: 'DELETE_SUCCESS' });
+      } catch (err) {
+        toast.error(getError(error));
+        dispatch({
+          type: 'DELETE_FAIL',
+        });
+      }
     }
   };
 
@@ -166,32 +214,39 @@ export default function TissuListPage() {
         </Col>
 
         <Col md={10} className="shadow p-5">
-          <Row className="align-items-between">
-            <Col md={4}>
-              <h1>Tissus</h1>
-            </Col>
-            <Col md={8}>
-              <Form onSubmit={submitHandler}>
-                <Row>
-                  <Col md={8}>
-                    <Form.Group controlId="name">
-                      <Form.Control
-                        placeholder="Nom"
-                        value={name}
-                        onChange={(e) => setName(e.target.value)}
-                        required
-                      />
-                    </Form.Group>
-                  </Col>
-                  <Col md={4}>
-                    <div className="mb-3">
-                      <Button type="submit">Ajouter</Button>
-                    </div>
-                  </Col>
-                </Row>
-              </Form>
-            </Col>
-          </Row>
+          <h1>Tissus</h1>
+          <Form onSubmit={submitHandler}>
+            <Row className="align-items-center">
+              <Col md={5} className="mt-2">
+                <Form.Group controlId="name">
+                  <Form.Control
+                    placeholder="Nom"
+                    value={name}
+                    onChange={(e) => setName(e.target.value)}
+                    required
+                  />
+                </Form.Group>
+              </Col>
+              <Col md={5} className="mt-2">
+                <Form.Group controlId="imageFile">
+                  <Form.Control
+                    type="file"
+                    ref={imageInputRef}
+                    onChange={uploadFileHandler}
+                  />
+                </Form.Group>
+              </Col>
+              <Col md={2} className="mt-2">
+                {loadingUpload ? (
+                  <LoadingBox />
+                ) : (
+                  <Button disabled={!name} type="submit" className="w-100">
+                    <FontAwesomeIcon icon={faPlus} />
+                  </Button>
+                )}
+              </Col>
+            </Row>
+          </Form>
 
           <hr />
           {loading ? (
@@ -203,6 +258,7 @@ export default function TissuListPage() {
               <thead>
                 <tr>
                   <th>Nom</th>
+                  <th>Image</th>
                   <th>Action</th>
                 </tr>
               </thead>
@@ -210,8 +266,14 @@ export default function TissuListPage() {
                 {tissus.map((tissu) => (
                   <tr key={tissu._id}>
                     <td>{tissu.name}</td>
+                    <td>
+                      {tissu.image && (
+                        <Image thumbnail src={tissu.image}></Image>
+                      )}
+                    </td>
 
                     <td>
+                      <ModalTissuEdit id={tissu._id} />
                       <Button
                         className="btn btn-sm"
                         type="button"
